Moved ViewParser to seperate file
