+++ src/utilities/ComponentClass.js
@@ -1,6 +1,7 @@
 import Radi from '../Radi';
 import { GLOBALS } from '../consts/GLOBALS';
 import { clone } from './clone';
+import ViewParser from './ViewParser';
 
 export class Component {
   constructor(o) {
@@ -10,9 +11,11 @@ export class Component {
       props: clone(o.props),
       actions: o.actions,
       view: o.view,
-      $view: o.$view,
       $mixins: this.$mixins || {},
     };
+
+    const parsedView = new ViewParser(this.o.view).parse();
+    this.o.$view = parsedView;
 
     this.$mixins = {};
   }
