+++ src/utilities/__tests__/ViewParser.test.js
@@ -1,0 +1,8 @@
+import ViewParser from '../ViewParser';
+
+// TODO: Test properly
+describe('ViewParser.js', () => {
+  it('doesn\'t crash', () => {
+    new ViewParser('test');
+  });
+});
